add: voteResult(미완), fix: createVote
